[#1] [suman/surajit] register as a player on game
What changes have you made?
changed the server.js and added package.json and for handle the server
portion routes.js added.
*

Why do we need this change?
for login the player and join to the game
*

What can this change impact?
Now a player abel to register in the game.
*