[tina/jishnu]-Disable & Enable player's hand according to their turn
What changes have you made?

We disable the player's hand when the player's turn is not.

Why do we need this change?

We need to stop play other player whose turn is not.

What can this change impact?

Now only whose turn is going on he can able to play cards.